last_move: Option<(usize, usize)>

+++ src/constants.rs
@@ -2,7 +2,7 @@ use nannou::color::Srgb;
 
 pub const BOARD_SIZE: usize = 19;
 pub const HALF_BOARD_SIZE: usize = BOARD_SIZE >> 1;
-pub const WINDOW_SIZE: usize = 800;
+pub const WINDOW_SIZE: u32 = 800;
 pub const WINDOW_MARGIN: f32 = 35.0;
 pub const CELL_SIZE: f32 = (WINDOW_SIZE as f32 - 2.0 * WINDOW_MARGIN) / (BOARD_SIZE - 1) as f32;
 pub const DOT_SPACING: usize = 6;
